Move condition to top-level directory
