+++ src/card-controller/types.ts
@@ -1,4 +1,5 @@
 import type { CameraManager } from '../camera-manager/manager';
+import type { ConditionStateManager } from '../conditions/state-manager';
 import type { Automation } from '../config/types';
 import type { EntityRegistryManager } from '../utils/ha/registry/entity';
 import type { ResolvedMediaCache } from '../utils/ha/resolved-media';
@@ -6,7 +7,6 @@ import type { ActionsManager } from './actions/actions-manager';
 import type { AutomationsManager } from './automations-manager';
 import type { CameraURLManager } from './camera-url-manager';
 import type { CardElementManager } from './card-element-manager';
-import type { ConditionStateManager } from './conditions/state-manager';
 import type { ConfigManager } from './config/config-manager';
 import type { DefaultManager } from './default-manager';
 import type { DownloadManager } from './download-manager';
