+++ src/card.ts
@@ -7,7 +7,6 @@ import { Ref, createRef, ref } from 'lit/directives/ref.js';
 import { styleMap } from 'lit/directives/style-map.js';
 import 'web-dialog';
 import { actionHandler } from './action-handler-directive.js';
-import { ConditionStateManagerGetEvent } from './card-controller/conditions/state-manager-via-event.js';
 import { CardController } from './card-controller/controller';
 import { MenuButtonController } from './components-lib/menu-button-controller';
 import './components/elements.js';
@@ -23,6 +22,7 @@ import './components/status-bar';
 import './components/thumbnail-carousel.js';
 import './components/views.js';
 import { AdvancedCameraCardViews } from './components/views.js';
+import { ConditionStateManagerGetEvent } from './conditions/state-manager-via-event.js';
 import {
   AdvancedCameraCardConfig,
   MenuItem,
