+++ src/conditions/state-manager-via-event.ts
