+++ src/conditions/state-manager.ts
